+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/util/ImageUtil.java
@@ -68,15 +68,19 @@ public class ImageUtil {
   }
 
   public static void drawTimestamp(File file) throws IOException {
+    drawWatermark(file, DateUtil.formatDateTime(new Date()), Color.RED);
+  }
+
+  public static void drawWatermark(File file, String watermark, Color color) throws IOException {
     BufferedImage bufferedImage = loadImage(file);
     if (bufferedImage != null) {
       Graphics g = bufferedImage.getGraphics();
       Graphics2D g2d = (Graphics2D) g.create();
-      g2d.setColor(Color.RED);
+      g2d.setColor(color);
       g2d.setFont(new Font("TimesRoman", Font.BOLD, 18));
-      g2d.drawString(DateUtil.formatDateTime(new Date()), 12, 30);
+      g2d.drawString(watermark, 12, 30);
       file.delete();
-      writeJPG(bufferedImage, file);
+      write(bufferedImage, file);
     }
   }