+++ RELEASE_NOTES.md
@@ -3,6 +3,9 @@
 ## Changes
 
 - **Preferences**: Fixed refresh of Avatar when uploading a new image
+- **DMD Position Tool**: 
+  - Show watermark when DMD image comes from frontend or PupPack
+  - Fixed message when B2S DMD is hidden, instead of uploading an Image, propose to show the image
 
 ---